A first approach at making grammar models be able to do semantics by the means of a delegate.
It needs much more work, and probably a refactoring of the whole thingamabop to have it make sense.  Maybe for a 2.0 version?

Parsing through the models might be extremely fast because of the reduced code footprint, and it could be done after grammars and semantics are debugged through the generated top down parsers.